Move zipFile method to src/util.js module. Fix issue with collecting examples keys. Change log messages. Disable files minification due to storage properties
