+++ src/tasks/process-examples.js
@@ -1,8 +1,6 @@
 'use strict';
 
-var fs = require('fs'),
-    zlib = require('zlib'),
-    path = require('path'),
+var path = require('path'),
     util = require('util'),
 
     vow = require('vow'),
@@ -26,46 +24,6 @@ function readFiles(baseDir) {
         err ? def.reject(err) : def.resolve(files);
     });
     return def.promise();
-}
-
-/**
- * Compress given file
- * @param {Target} target - target object
- * @param {String} filePath - path to source file
- * @returns {*}
- */
-function zipFile(target, filePath) {
-    var basePath = target.getTempPath(),
-        sPath = path.join(basePath, filePath),
-        dPath = sPath + '.zip';
-
-    return vowFs.isSymLink(sPath)
-        .then(function (isSymlink) {
-            if (isSymlink) {
-                console.log('find symlink %s', filePath);
-                return vow.resolve();
-            }
-
-            var def = vow.defer(),
-                readStream = fs.createReadStream(sPath),
-                writeStream = fs.createWriteStream(dPath);
-
-            readStream
-                .pipe(zlib.Gzip())
-                .pipe(writeStream)
-                .on('error', function (err) {
-                    logger.warn(util.format('error occur while compressing: %s', filePath), module);
-                    def.reject(err);
-                })
-                .on('close', function () {
-                    logger.verbose(util.format('compressed file: %s', filePath), module);
-                    fs.rename(dPath, sPath, function (err) {
-                        err ? def.reject(err) : def.resolve(filePath);
-                    });
-                });
-
-            return def.promise();
-    });
 }
 
 /**
@@ -108,24 +66,42 @@ module.exports = function (target) {
 
     return readFiles(target.getTempPath())
         .then(function (files) {
+            files = files.filter(function (file) {
+                if (file.match(/README\.md$/)) {
+                    return false;
+                }
+                if (file.match(/desktop\.sets\/(\.bem|catalogue|index|jscatalogue)/)) {
+                    return false;
+                }
+                if (file.match(/\/\.bem\//)) {
+                    return false;
+                }
+                if (file.match(/data\.json$/)) {
+                    return false;
+                }
+                return true;
+            });
+
             var portions = utility.separateArrayOnChunks(files, openFilesLimit);
 
             logger.debug(util.format('example files count: %s', files.length), module);
-            logger.debug(util.format('compression will be executed in %s steps', portions.length), module);
+            logger.debug(util.format('processing will be executed in %s steps', portions.length), module);
 
             return portions.reduce(function (prev, item, index) {
                 prev = prev.then(function () {
-                    logger.debug(util.format('compress and send files in range %s - %s',
+                    logger.debug(util.format('send files in range %s - %s',
                         index * openFilesLimit, (index + 1) * openFilesLimit), module);
 
                     var promises = item.map(function (_item) {
-                        return zipFile(target, _item).then(function () {
+                        // TODO temporary disable compression
+                        // return utility.zipFile(path.join(target.getTempPath(), _item)).then(function () {
                             return sendToStorage(target, _item);
-                        });
+                        // });
                     });
 
                     return vow.all(promises).then(function (keys) {
-                        exampleKeys.concat(keys);
+                        exampleKeys = exampleKeys.concat(keys);
+                        return exampleKeys;
                     });
                 });
                 return prev;
